Enable runtime backend selection
Now that there are two functional compositor backends, it makes sense to
allow for them to coexist in a single build. If they are both built in by
default it is more likely that we will catch build breakages sooner, and it
allows us to have a fallback backend in production.

We are currently exposing three classes per backend, which creates a
challenge as it makes sense for a backend to be in charge of probing for
support, but it should be all-or-nothing, or else we might end up with mixed
objects.

This change explicitly groups factory functions for each backend and binds a
grouping if it passes a call to is_supported() for the respective backend.
is_supported() simply checks if a context can be created and initialized. In
the future we can expand this definition to be more strict about what
hardware/software capabilities are required

Jira: IAHWC-40
Test: Both backends are still working with demo apps

Signed-off-by: Kevin Strasser <kevin.strasser@intel.com>